+++ Services/Lts/V2/Model/UpdateAccessConfigRequestBody.cs
@@ -52,6 +52,12 @@ namespace HuaweiCloud.SDK.Lts.V2.Model
         [JsonProperty("binary_collect", NullValueHandling = NullValueHandling.Ignore)]
         public bool? BinaryCollect { get; set; }
 
+        /// <summary>
+        /// CCE集群ID，CCE类型时，为必填
+        /// </summary>
+        [JsonProperty("cluster_id", NullValueHandling = NullValueHandling.Ignore)]
+        public string ClusterId { get; set; }
+
 
 
         /// <summary>
@@ -67,6 +73,7 @@ namespace HuaweiCloud.SDK.Lts.V2.Model
             sb.Append("  accessConfigTag: ").Append(AccessConfigTag).Append("\n");
             sb.Append("  logSplit: ").Append(LogSplit).Append("\n");
             sb.Append("  binaryCollect: ").Append(BinaryCollect).Append("\n");
+            sb.Append("  clusterId: ").Append(ClusterId).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
         }
@@ -118,6 +125,11 @@ namespace HuaweiCloud.SDK.Lts.V2.Model
                     this.BinaryCollect == input.BinaryCollect ||
                     (this.BinaryCollect != null &&
                     this.BinaryCollect.Equals(input.BinaryCollect))
+                ) && 
+                (
+                    this.ClusterId == input.ClusterId ||
+                    (this.ClusterId != null &&
+                    this.ClusterId.Equals(input.ClusterId))
                 );
         }
 
@@ -141,6 +153,8 @@ namespace HuaweiCloud.SDK.Lts.V2.Model
                     hashCode = hashCode * 59 + this.LogSplit.GetHashCode();
                 if (this.BinaryCollect != null)
                     hashCode = hashCode * 59 + this.BinaryCollect.GetHashCode();
+                if (this.ClusterId != null)
+                    hashCode = hashCode * 59 + this.ClusterId.GetHashCode();
                 return hashCode;
             }
         }